update size create

+++ src/controllers/category.controller.js
@@ -3,16 +3,17 @@ import { uploadImage } from "../utils/cloudinary.js";
 import * as fs from "fs/promises";
 
 export const categoryController = {
-
   create: async (req, res) => {
-
     const image = [];
+    const { name } = req.body;
 
     try {
-      const { name } = req.body;
+      if (!name) {
+        return res.status(400).json({ message: "El nombre es obligatorio" });
+      }
+
       for (let i = 0; i < req.files.length; i++) {
         const result = await uploadImage(req.files[i].path);
-        console.log(result);
         const url = result.secure_url;
         image.push(url);
         await fs.unlink(req.files[i].path);
@@ -20,13 +21,15 @@ export const categoryController = {
 
       const category = new Category({
         name,
-        image
+        image,
       });
       await category.save();
-      res.json({message: 'create category'})
+      return res.status(201).json({ message: "Categoria creada"});
     } catch (error) {
-      console.log(error);
-        res.json({message: "Duplicate name in categories"})
+      if (error.code === 11000) {
+        return res.status(409).json({ message: "Categoria ya existe" });
+      }
+      return res.status(500).json({ message: error.message });
     }
   },
 
@@ -48,11 +51,11 @@ export const categoryController = {
   },
 
   delete: async (req, res) => {
-    const { _id } = req.params
+    const { _id } = req.params;
     try {
-      const category = await Category.findOne({_id});
+      const category = await Category.findOne({ _id });
       category.set({
-        status: false
+        status: false,
       });
       await category.save();
       res.json(category);
@@ -62,7 +65,7 @@ export const categoryController = {
   },
 
   deleteAll: async (req, res) => {
-    await Category.deleteMany()
-    return res.json({message: "deleted categories"})
-  }
+    await Category.deleteMany();
+    return res.json({ message: "deleted categories" });
+  },
 };
